+++ drape/cpu_buffer.hpp
@@ -12,15 +12,15 @@ class CPUBuffer : public BufferBase
 {
   typedef BufferBase TBase;
 public:
-  CPUBuffer(uint8_t elementSize, uint16_t capacity);
+  CPUBuffer(uint8_t elementSize, uint32_t capacity);
   ~CPUBuffer();
 
-  void UploadData(void const * data, uint16_t elementCount);
+  void UploadData(void const * data, uint32_t elementCount);
   // Set memory cursor on element with number == "elementNumber"
   // Element numbers start from 0
-  void Seek(uint16_t elementNumber);
+  void Seek(uint32_t elementNumber);
   // Check function. In real world use must use it only in assert
-  uint16_t GetCurrentElementNumber() const;
+  uint32_t GetCurrentElementNumber() const;
   unsigned char const * Data() const;
 
 private:
