+++ drape/data_buffer.cpp
@@ -4,7 +4,7 @@
 namespace dp
 {
 
-DataBuffer::DataBuffer(uint8_t elementSize, uint16_t capacity)
+DataBuffer::DataBuffer(uint8_t elementSize, uint32_t capacity)
   : m_impl(make_unique_dp<CpuBufferImpl>(elementSize, capacity))
 {
 }
@@ -18,7 +18,7 @@ ref_ptr<DataBufferBase> DataBuffer::GetBuffer() const
 void DataBuffer::MoveToGPU(GPUBuffer::Target target)
 {
   // if currentSize is 0 buffer hasn't been filled on preparation stage, let it be filled further
-  uint16_t const currentSize = m_impl->GetCurrentSize();
+  uint32_t const currentSize = m_impl->GetCurrentSize();
   if (currentSize != 0)
     m_impl = make_unique_dp<GpuBufferImpl>(target, m_impl->Data(),m_impl->GetElementSize(), currentSize);
   else
@@ -38,7 +38,7 @@ DataBufferMapper::~DataBufferMapper()
   m_buffer->GetBuffer()->Unmap();
 }
 
-void DataBufferMapper::UpdateData(void const * data, uint16_t elementOffset, uint16_t elementCount)
+void DataBufferMapper::UpdateData(void const * data, uint32_t elementOffset, uint32_t elementCount)
 {
   m_buffer->GetBuffer()->UpdateData(m_ptr, data, elementOffset, elementCount);
 }
