+++ drape/data_buffer.hpp
@@ -11,15 +11,15 @@ class DataBufferBase
 public:
   virtual ~DataBufferBase() {}
 
-  virtual uint16_t GetCapacity() const = 0;
-  virtual uint16_t GetCurrentSize() const = 0;
-  virtual uint16_t GetAvailableSize() const = 0;
+  virtual uint32_t GetCapacity() const = 0;
+  virtual uint32_t GetCurrentSize() const = 0;
+  virtual uint32_t GetAvailableSize() const = 0;
   virtual uint8_t GetElementSize() const = 0;
-  virtual void Seek(uint16_t elementNumber) = 0;
+  virtual void Seek(uint32_t elementNumber) = 0;
   virtual void const * Data() const = 0;
 
-  virtual void UploadData(void const * data, uint16_t elementCount) = 0;
-  virtual void UpdateData(void * destPtr, void const * srcPtr, uint16_t elementOffset, uint16_t elementCount) = 0;
+  virtual void UploadData(void const * data, uint32_t elementCount) = 0;
+  virtual void UpdateData(void * destPtr, void const * srcPtr, uint32_t elementOffset, uint32_t elementCount) = 0;
 
   virtual void Bind() = 0;
   virtual void * Map() = 0;
@@ -30,7 +30,7 @@ public:
 class DataBuffer
 {
 public:
-  DataBuffer(uint8_t elementSize, uint16_t capacity);
+  DataBuffer(uint8_t elementSize, uint32_t capacity);
 
   ref_ptr<DataBufferBase> GetBuffer() const;
   void MoveToGPU(GPUBuffer::Target target);
@@ -46,7 +46,7 @@ public:
   DataBufferMapper(ref_ptr<DataBuffer> buffer);
   ~DataBufferMapper();
 
-  void UpdateData(void const * data, uint16_t elementOffset, uint16_t elementCount);
+  void UpdateData(void const * data, uint32_t elementOffset, uint32_t elementCount);
 
 private:
   ref_ptr<DataBuffer> m_buffer;
