+++ drape/gpu_buffer.hpp
@@ -17,19 +17,19 @@ public:
   };
 
 public:
-  GPUBuffer(Target t, void const * data, uint8_t elementSize, uint16_t capacity);
+  GPUBuffer(Target t, void const * data, uint8_t elementSize, uint32_t capacity);
   ~GPUBuffer();
 
-  void UploadData(void const * data, uint16_t elementCount);
+  void UploadData(void const * data, uint32_t elementCount);
   void Bind();
 
   void * Map();
-  void UpdateData(void * gpuPtr, void const * data, uint16_t elementOffset, uint16_t elementCount);
+  void UpdateData(void * gpuPtr, void const * data, uint32_t elementOffset, uint32_t elementCount);
   void Unmap();
 
 protected:
   /// discard old data
-  void Resize(void const * data, uint16_t elementCount);
+  void Resize(void const * data, uint32_t elementCount);
 
 private:
   friend class GPUBufferMapper;
